# perf(sql): avoid full table scan in ASOF and LT JOIN with no additional columns