--- a/core/src/main/java/io/questdb/cairo/CairoConfiguration.java
+++ b/core/src/main/java/io/questdb/cairo/CairoConfiguration.java
@@ -346,7 +346,7 @@ public interface CairoConfiguration {
 
     long getSpinLockTimeout();
 
-    int getSqlAsOfJoinLookahead();
+    int getSqlAsOfJoinLookAhead();
 
     int getSqlCharacterStoreCapacity();