--- a/core/src/main/java/io/questdb/cairo/CairoConfigurationWrapper.java
+++ b/core/src/main/java/io/questdb/cairo/CairoConfigurationWrapper.java
@@ -542,6 +542,11 @@ public class CairoConfigurationWrapper implements CairoConfiguration {
     }
 
     @Override
+    public int getSqlAsOfJoinLookahead() {
+        return delegate.getSqlAsOfJoinLookahead();
+    }
+
+    @Override
     public int getSqlCharacterStoreCapacity() {
         return delegate.getSqlCharacterStoreCapacity();
     }