--- a/core/src/main/java/io/questdb/cairo/CairoConfigurationWrapper.java
+++ b/core/src/main/java/io/questdb/cairo/CairoConfigurationWrapper.java
@@ -542,8 +542,8 @@ public class CairoConfigurationWrapper implements CairoConfiguration {
     }
 
     @Override
-    public int getSqlAsOfJoinLookahead() {
-        return delegate.getSqlAsOfJoinLookahead();
+    public int getSqlAsOfJoinLookAhead() {
+        return delegate.getSqlAsOfJoinLookAhead();
     }
 
     @Override