--- a/core/src/main/java/io/questdb/griffin/engine/join/AsOfJoinFastNoKeyRecordCursorFactory.java
+++ b/core/src/main/java/io/questdb/griffin/engine/join/AsOfJoinFastNoKeyRecordCursorFactory.java
@@ -81,7 +81,7 @@ public class AsOfJoinFastNoKeyRecordCursorFactory extends AbstractJoinRecordCurs
 
     @Override
     public void toPlan(PlanSink sink) {
-        sink.type("AsOf Join");
+        sink.type("AsOf Join Fast Scan");
         sink.child(masterFactory);
         sink.child(slaveFactory);
     }