--- a/core/src/main/java/io/questdb/griffin/engine/join/AsOfJoinNoKeyFastRecordCursorFactory.java
+++ b/core/src/main/java/io/questdb/griffin/engine/join/AsOfJoinNoKeyFastRecordCursorFactory.java
@@ -271,6 +271,7 @@ public class AsOfJoinNoKeyFastRecordCursorFactory extends AbstractJoinRecordCurs
                     }
                     if (slaveFrameRow < frame.getRowHi()) {
                         // Fallback to binary search.
+                        // Find the last value less or equal to the master timestamp.
                         long foundRow = binarySearch(
                                 masterTimestamp,
                                 slaveFrameRow,