Closes #4182

In the case of trivial ASOF/LT JOIN queries, we may have scanned the right-hand table from the oldest partition to the ones required by left-hand table timestamps. For large tables, this led to lots of redundant disk reads.

Problematic query examples:
```sql
SELECT * FROM a ASOF JOIN b WHERE a.ts IN '2024-02';

SELECT * FROM a LT JOIN b WHERE a.ts IN '2024-02';
```

This PR fixes that by adding new `Fast*` record cursor factories for non-keyed ASOF/LT JOINs. These factories use new `TimeFrameRecordCursor`/`TimeFrame` interfaces. These interfaces allow factories to scan partitions lazily, i.e. without entering them, and access rows randomly, when a partition is entered. The time frame cursor is only available for trivial single-table queries/factories.

The factories scan time frames and, once a suitable frame is found, employ a linear scan of a small number of rows followed by a binary search. The linear scan is necessary to optimize for worst-case scenarios like self-join or join of identical tables.

New factories can be seen in `EXPLAIN` output as `AsOf/Lt Join Fast Scan`.

Also includes:
* Simplify `ASOF/LT JOIN b ON (ts)` to `ASOF/LT JOIN b`. Previously we were treating these joins as keyed.

## Benchmarks

Held on my machine, Ubuntu 22.04, i7-1185G7, 32GB RAM, NVMe SSD.

DDLs:
```sql
create table a (seq int, ts timestamp) timestamp(ts) partition by DAY;
create table b (seq int, ts timestamp) timestamp(ts) partition by DAY;

insert into a select * from (select rnd_int() seq,
    timestamp_sequence(to_timestamp('2024-02-01T01:01:01', 'yyyy-MM-ddTHH:mm:ss'),
    3600000L) ts from long_sequence(10000)) timestamp (ts);

insert into b select * from (select rnd_int() seq, timestamp_sequence(
    to_timestamp('2023-02-01T00:00:00', 'yyyy-MM-ddTHH:mm:ss'),
    3600000L) ts from long_sequence(10000000)) timestamp (ts);
```

Best case (right hand has lots of partitions prior to the required one):
```sql
-- master: 485ms (cold), 130ms (hot)
-- this patch: 94ms (cold), 34ms (hot)
SELECT * FROM a ASOF JOIN b WHERE a.ts BETWEEN '2024-02-01T00:00:00' AND '2024-03-11T00:00:00';

-- master: 120ms (hot)
-- this patch: 35ms (hot)
SELECT * FROM a LT JOIN b WHERE a.ts BETWEEN '2024-02-01T00:00:00' AND '2024-03-11T00:00:00';
```

Worst case (self-join):
```sql
-- master: 35ms (hot)
-- this patch: 38ms (hot)
SELECT * FROM b b1 ASOF JOIN b b2 ON(ts) LIMIT 10000;

-- master: 26ms (hot)
-- this patch: 31ms (hot)
SELECT * FROM b b1 LT JOIN b b2 ON(ts) LIMIT 10000;
```