@ideoma 2024-02-09
#### [PR Coverage check]

:heart_eyes: **pass** : 450 / 470 (95.74%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/cairo/FullBwdDataFrameCursorFactory.java|0|1|00.00%|
|:large_blue_circle:|io/questdb/cairo/IntervalBwdDataFrameCursorFactory.java|0|1|00.00%|
|:large_blue_circle:|io/questdb/cairo/sql/RecordCursorFactory.java|1|2|50.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/DataFrameRecordCursorFactory.java|11|12|91.67%|
|:large_blue_circle:|io/questdb/griffin/engine/join/AsOfJoinNoKeyFastRecordCursorFactory.java|155|162|95.68%|
|:large_blue_circle:|io/questdb/griffin/engine/join/LtJoinNoKeyFastRecordCursorFactory.java|158|166|95.18%|
|:large_blue_circle:|io/questdb/griffin/engine/table/TableReaderTimeFrameCursor.java|80|81|98.77%|
|:large_blue_circle:|io/questdb/PropServerConfiguration.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/cairo/IntervalFwdDataFrameCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/table/TouchTableFunctionFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/cairo/FullFwdDataFrameCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/cairo/DefaultCairoConfiguration.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/SelectedRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/RegisteredRecordCursorFactory.java|3|3|100.00%|
|:large_blue_circle:|io/questdb/PropertyKey.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/SqlCodeGenerator.java|30|30|100.00%|
|:large_blue_circle:|io/questdb/cairo/BinarySearch.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/cairo/CairoConfigurationWrapper.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/std/Rows.java|1|1|100.00%|